backend/requirements.txt
# AGI System Backend Dependencies
# Python 3.11+

# ============================================================================
# WEB FRAMEWORK
# ============================================================================
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# ============================================================================
# LANGCHAIN ECOSYSTEM
# ============================================================================
langchain>=0.1.4
langchain-core>=0.1.32
langchain-community>=0.0.17
langgraph>=0.0.20
langsmith>=0.0.83

# LangChain integrations
langchain-anthropic==0.1.1
# langchain-postgres==0.0.2
langchain-cohere==0.1.0
langchain-voyageai==0.1.0

# ============================================================================
# GRAPHQL
# ============================================================================
strawberry-graphql[fastapi]==0.219.1

# ============================================================================
# DATABASE
# ============================================================================
asyncpg==0.29.0
psycopg2-binary==2.9.9
pgvector==0.2.4
sqlalchemy==2.0.25
alembic==1.13.1

# ============================================================================
# REDIS
# ============================================================================
redis==5.0.1

# ============================================================================
# VECTOR & EMBEDDINGS
# ============================================================================
voyageai==0.2.1
cohere>=5.1.4
sentence-transformers==2.3.1
faiss-cpu==1.7.4
scikit-learn==1.4.0
numpy==1.24.3

# ============================================================================
# ERROR TRACKING & OBSERVABILITY
# ============================================================================
sentry-sdk[fastapi]==1.39.2
prometheus-client==0.19.0

# ============================================================================
# UTILITIES
# ============================================================================
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
tenacity==8.2.3
httpx==0.26.0

# ============================================================================
# TESTING
# ============================================================================
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
